Refactor/ 11.29 리팩토링 완
