+++ src/components/COMMON/AlertIndicator.tsx
@@ -1,3 +1,4 @@
+//  경고창을 띄웁니다
 const AlertIndicator = ({ children }: { children: React.ReactNode }) => {
   return (
     <div
