+++ src/components/COMMON/Header.tsx
@@ -1,53 +1,15 @@
-import { useNavigate } from 'react-router-dom';
-import { signOut } from '../../firebase/userManage';
-import { auth } from '../../firebase/firebase';
 import AdminController from './AdminController';
-import { HK_USER } from '../../static/const/variable';
-import { browserSessionPersistence, onAuthStateChanged, setPersistence } from 'firebase/auth';
-import { useEffect } from 'react';
-import { Category, User } from '../../static/const/type';
-import { useRecoilState } from 'recoil';
-import { currentUserState } from '../../atom/currentUserState';
-import { currentCategory } from '../../atom/currentCategory';
+import { useSetRecoilState } from 'recoil';
+import { category, currentCategory } from '../../atom/currentCategory';
 import SearchBox from '../SearchPage/SearchBox';
+import { useNavigate } from 'react-router-dom';
+import LoginInfoBox from './LoginInfoBox';
 
 type Props = {};
 
 const Header = (props: Props) => {
   const navi = useNavigate();
-
-  const [currentUser, setCurrentUser] = useRecoilState(currentUserState);
-
-  const [_, setCategory] = useRecoilState(currentCategory);
-
-  useEffect(() => {
-    // 인증 상태 지속성을 세션으로 설정합니다.
-    setPersistence(auth, browserSessionPersistence)
-      .then(() => {
-        return onAuthStateChanged(auth, (user) => {
-          if (user) {
-            const curretUser = {
-              uid: user.uid,
-              displayName: user.displayName,
-              email: user.email,
-              created_at: user.metadata.creationTime
-            } as User;
-
-            sessionStorage.setItem(HK_USER, JSON.stringify(curretUser));
-            setCurrentUser(curretUser);
-            // console.log('User is logged in:', user);
-          } else {
-            sessionStorage.removeItem(HK_USER);
-            setCurrentUser(null);
-            // console.log('User is logged out:', user);
-          }
-        });
-      })
-      .catch((error) => {
-        // Handle errors
-        console.error('Error setting persistence:', error);
-      });
-  }, []);
+  const setCategory = useSetRecoilState(currentCategory);
 
   return (
     <>
@@ -72,59 +34,24 @@ const Header = (props: Props) => {
                 홈
               </a>
             </li>
-            {/* <li>
-              <button onClick={() => navi('/product/1')}>상품페이지로</button>
-            </li> */}
           </ul>
 
           {/* 상품 카테고리 메뉴 */}
           <ul className="flex justify-between space-x-4 bg-gray-700">
-            <li>
-              <button className="hover:text-gray-300" onClick={() => setCategory(Category.ALL)}>
-                전체
-              </button>
-            </li>
-            <li>
-              <button className="hover:text-gray-300" onClick={() => setCategory(Category.HAT)}>
-                모자
-              </button>
-            </li>
-            <li>
-              <button className="hover:text-gray-300" onClick={() => setCategory(Category.T_SHIRTS)}>
-                티셔츠
-              </button>
-            </li>
-            <li>
-              <button className="hover:text-gray-300" onClick={() => setCategory(Category.TRAINING_CLOTHS)}>
-                트레이닝복
-              </button>
-            </li>
+            {Object.entries(category)
+              .map(([id, name]) => ({ id, name }))
+              .map((val, idx) => {
+                return (
+                  <li key={idx}>
+                    <button className="hover:text-gray-300" onClick={() => setCategory(val.name)}>
+                      {val.name}
+                    </button>
+                  </li>
+                );
+              })}
           </ul>
 
-          <div>
-            <nav className="bg-gray-600 text-white text-end px-5 my">
-              {currentUser ? (
-                <>
-                  {currentUser.email}님, 환영합니다!{' '}
-                  <div className="flex justify-end">
-                    <button className="text-red ms-2 bg-gray-500" onClick={() => navi('/mypage/1')}>
-                      마이페이지
-                    </button>
-                    <form onSubmit={signOut}>
-                      <button className="text-red ms-2 bg-red-800">로그아웃</button>
-                    </form>
-                  </div>
-                </>
-              ) : (
-                <>
-                  로그인 정보 없음
-                  <button className="text-blue ms-3" onClick={() => navi('/login')}>
-                    로그인
-                  </button>
-                </>
-              )}
-            </nav>
-          </div>
+          <LoginInfoBox />
         </nav>
       </header>
     </>
