+++ src/components/COMMON/ProductCard.tsx
@@ -12,17 +12,9 @@ const ProductCard = ({ item }: Props) => {
   const { deleteProductMutation } = useProductQuery();
 
   return (
-    <Link
-      to={`/products/${item.id}`}
-      state={item}
-      className="mx-3 mt-6 flex flex-col rounded-lg bg-white shadow-[0_2px_15px_-3px_rgba(0,0,0,0.07),0_10px_20px_-2px_rgba(0,0,0,0.04)] dark:bg-neutral-700 sm:shrink-0 sm:grow sm:basis-0 hover:cursor-pointer hover:border-black border-2"
-    >
+    <Link to={`/products/${item.id}`} state={item} className="productCardContainer">
       <div className="overflow-hidden rounded-t-lg">
-        <img
-          className="hover:scale-125 transition-transform duration-300"
-          src={item.imgs[0]}
-          alt="Hollywood Sign on The Hill"
-        />
+        <img className="hover:scale-125 transition-transform duration-300" src={item.imgs[0]} alt="이미지없음" />
       </div>
       <div className="p-6">
         <h5 className="mb-2 text-xl font-medium leading-tight text-neutral-800 dark:text-neutral-50">{item.name}</h5>
