+++ src/pages/Login.tsx
@@ -1,0 +1,116 @@
+import React, { useState } from 'react';
+
+import { useNavigate } from 'react-router-dom';
+import { signInWithEmailAndPassword } from 'firebase/auth';
+import { auth } from '../firebase/firebase';
+import FacebookLoginButton from '../components/Login/FacebookLoginButton';
+import GoogleLoginButton from '../components/Login/GoogleLoginButton';
+import InputField from '../components/COMMON/InputField';
+
+type Props = {};
+
+const Login = (props: Props) => {
+  const [email, setEmail] = useState('');
+  const [pass, setPass] = useState('');
+  const navi = useNavigate();
+
+  const emailHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setEmail(e.target.value);
+  };
+  const passHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setPass(e.target.value);
+  };
+  const confirmLogin = async (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    try {
+      const userCredential = await signInWithEmailAndPassword(auth, email, pass);
+
+      // console.log(userCredential);
+      const user = userCredential.user;
+      // console.log('User signed in:', user);
+      alert('로그인 완료!');
+      navi('/');
+    } catch (error) {
+      alert('로그인 실패!');
+      console.error('Error signing in:', error);
+      return;
+    }
+  };
+
+  return (
+    <section className="h-screen">
+      <div className="h-full">
+        <div className="g-6 flex h-full flex-wrap items-center justify-center lg:justify-between">
+          <div className="shrink-1 mb-12 grow-0 basis-auto md:mb-0 md:w-9/12 md:shrink-0 lg:w-6/12 xl:w-6/12">
+            <img
+              src="https://tecdn.b-cdn.net/img/Photos/new-templates/bootstrap-login-form/draw2.webp"
+              className="w-full"
+              alt="Sample image"
+            />
+          </div>
+
+          <div className="mb-12 md:mb-0 md:w-8/12 lg:w-5/12 xl:w-5/12">
+            <form onSubmit={confirmLogin}>
+              <div className="flex flex-row items-center justify-center lg:justify-start">
+                <p className="mb-0 mr-4 text-lg">Sign in with</p>
+                <FacebookLoginButton />
+
+                <GoogleLoginButton />
+              </div>
+
+              <div className="my-4 flex items-center before:mt-0.5 before:flex-1 before:border-t before:border-neutral-300 after:mt-0.5 after:flex-1 after:border-t after:border-neutral-300">
+                <p className="mx-4 mb-0 text-center font-semibold dark:text-white">Or</p>
+              </div>
+
+              {/* 이메일 입력란 */}
+              <InputField
+                label="이메일"
+                type="email"
+                value={email}
+                onChange={emailHandler}
+                placeholder="이메일"
+                required
+              />
+              {/* 패스워드 입력란 */}
+              <InputField
+                label="패스워드"
+                type="password"
+                value={pass}
+                onChange={passHandler}
+                placeholder="••••••"
+                required
+              />
+
+              <div className="mb-6 flex items-center justify-between">
+                <div className="mb-[0.125rem] block min-h-[1.5rem] pl-[1.5rem]">
+                  <input
+                    className="relative float-left -ml-[1.5rem] mr-[6px] mt-[0.15rem] h-[1.125rem] w-[1.125rem] appearance-none rounded-[0.25rem] border-[0.125rem] border-solid border-neutral-300 outline-none before:pointer-events-none before:absolute before:h-[0.875rem] before:w-[0.875rem] before:scale-0 before:rounded-full before:bg-transparent before:opacity-0 before:shadow-[0px_0px_0px_13px_transparent] before:content-[''] checked:border-primary checked:bg-primary checked:before:opacity-[0.16] checked:after:absolute checked:after:-mt-px checked:after:ml-[0.25rem] checked:after:block checked:after:h-[0.8125rem] checked:after:w-[0.375rem] checked:after:rotate-45 checked:after:border-[0.125rem] checked:after:border-l-0 checked:after:border-t-0 checked:after:border-solid checked:after:border-white checked:after:bg-transparent checked:after:content-[''] hover:cursor-pointer hover:before:opacity-[0.04] hover:before:shadow-[0px_0px_0px_13px_rgba(0,0,0,0.6)] focus:shadow-none focus:transition-[border-color_0.2s] focus:before:scale-100 focus:before:opacity-[0.12] focus:before:shadow-[0px_0px_0px_13px_rgba(0,0,0,0.6)] focus:before:transition-[box-shadow_0.2s,transform_0.2s] focus:after:absolute focus:after:z-[1] focus:after:block focus:after:h-[0.875rem] focus:after:w-[0.875rem] focus:after:rounded-[0.125rem] focus:after:content-[''] checked:focus:before:scale-100 checked:focus:before:shadow-[0px_0px_0px_13px_#3b71ca] checked:focus:before:transition-[box-shadow_0.2s,transform_0.2s] checked:focus:after:-mt-px checked:focus:after:ml-[0.25rem] checked:focus:after:h-[0.8125rem] checked:focus:after:w-[0.375rem] checked:focus:after:rotate-45 checked:focus:after:rounded-none checked:focus:after:border-[0.125rem] checked:focus:after:border-l-0 checked:focus:after:border-t-0 checked:focus:after:border-solid checked:focus:after:border-white checked:focus:after:bg-transparent dark:border-neutral-600 dark:checked:border-primary dark:checked:bg-primary dark:focus:before:shadow-[0px_0px_0px_13px_rgba(255,255,255,0.4)] dark:checked:focus:before:shadow-[0px_0px_0px_13px_#3b71ca]"
+                    type="checkbox"
+                    value=""
+                    id="exampleCheck2"
+                  />
+                  <label className="inline-block pl-[0.15rem] hover:cursor-pointer">Remember me</label>
+                </div>
+
+                <a href="#!">Forgot password?</a>
+              </div>
+
+              <div className="text-center lg:text-left">
+                <button className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded">로그인</button>
+
+                <p className="mb-0 mt-2 pt-1 text-sm font-semibold">
+                  계정이 없으신가요?
+                  <a href="/register" className="font-medium text-blue-600 dark:text-blue-500 hover:underline">
+                    회원가입
+                  </a>
+                </p>
+              </div>
+            </form>
+          </div>
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default Login;
